[Fix #265] RoleModal.tsx 로직 수정
